requirements.txt
fastapi
uvicorn
tensorflow
pydantic
python-dotenv
pillow
